+++ vss_cli/rel_opts.py
@@ -40,13 +40,6 @@ page_opt = click.option(
     help='page results in a less-like format',
     show_default=True,
 )
-wait_opt = click.option(
-    '--wait/--no-wait',
-    is_flag=True,
-    help='wait for request(s) to complete',
-    default=False,
-    show_default=True,
-)
 dry_run_opt = click.option('--dry-run', is_flag=True, help='no changes')
 max_del_opt = click.option(
     '-m',